Updated the image layout
I've changed the image layout to allow for users to read the text from the book on hover instead of having the load the image up in a lightbox/new tab.

This did however result in the fact that only one column of images can be used at a time.